:truck: rename: Move Common package to root
