+++ src/main/kotlin/com/team1/mvp_test/common/error/EnterpriseErrorMessage.kt
@@ -1,4 +1,4 @@
-package com.team1.mvp_test.domain.common.error
+package com.team1.mvp_test.common.error
 
 enum class EnterpriseErrorMessage(val message: String) {
     ALREADY_EXISTS("이미 존재하는 기업입니다")
