+++ src/main/kotlin/com/team1/mvp_test/common/error/MemberErrorMessage.kt
@@ -1,4 +1,4 @@
-package com.team1.mvp_test.domain.common.error
+package com.team1.mvp_test.common.error
 
 enum class MemberErrorMessage(val message: String) {
     EMAIL_ALREADY_IN_USE("이미 사용중인 이메일입니다."),
