+++ src/main/kotlin/com/team1/mvp_test/common/exception/PasswordIncorrectException.kt
@@ -1,4 +1,4 @@
-package com.team1.mvp_test.domain.common.exception
+package com.team1.mvp_test.common.exception
 
 data class PasswordIncorrectException(
     override val message: String? = "Incorrect Password",
