+++ src/main/kotlin/com/team1/mvp_test/common/exception/dto/ErrorResponse.kt
@@ -1,0 +1,5 @@
+package com.team1.mvp_test.common.exception.dto
+
+data class ErrorResponse(
+    val message: String?
+)
