--- src/main/kotlin/com/team1/mvp_test/domain/common/exception/dto/ErrorResponse.kt
@@ -1,5 +1,0 @@
-package com.team1.mvp_test.domain.common.exception.dto
-
-data class ErrorResponse(
-    val message: String?
-)
